Time: 35 ms (42.96%) | Memory: 28.8 MB (91.70%) - LeetSync

+++ 633-sum-of-square-numbers/sum-of-square-numbers.cs
@@ -29,7 +29,7 @@ public class Solution {
         
         while (left<=right){
             
-           long current = left*left + (long)right*right;
+           long current = left*left + right*right;
             
             if (current > c){
                 right--;
